Added messages design and functionality and send message design

+++ SummitPower/src/main/java/com/iub/summitpower/features/senior_executive/services/SeniorExecutiveServicesImpl.java
@@ -216,6 +216,10 @@ public class SeniorExecutiveServicesImpl extends RepositoryUtils implements ISen
 
         switch (userType) {
             case SENIOR_EXECUTIVE:
+                if(Setup.currentUser.getUserType() == HR_EXECUTIVE) {
+                    ViewControlUtils.showAlert("You can not remove an Executive from the company.");
+                    return false;
+                }
                 return seniorExecutiveRepository.delete(userName);
 
             case HR_EXECUTIVE:
